refactor: treat realtime files as standard and get rid of old ones
